+++ symbol-processor-common/build.gradle
@@ -11,8 +11,6 @@ dependencies {
     api libs.ksp.api
     api libs.kotlin.reflect
 
-    implementation libs.slf4j.api
-
     testFixturesApi project(':aop:aop-symbol-processor')
     testFixturesApi libs.kotlin.reflect
